V2.6 change slave port camera file empty

+++ CQ33_RELI/lib/cq_file/camera/monitor_camera.sh
@@ -25,8 +25,8 @@ function monitor_cam_excute()
 
 function monitor_cam_result()
 {
-        local cam_excute=${log_dir}/camera_excute.log
-        local cam_statis=${log_dir}/camera_result.log
+	local cam_excute=${log_dir}/camera_excute.log
+	local cam_statis=${log_dir}/camera_result.log
 	local cycle_time=$(jq -r '.cycle_times' ${config_json})
 	local delay_time=$(jq -r '.delay_time' ${config_json})
 	local i
@@ -43,29 +43,35 @@ function monitor_cam_result()
                 } | tee ${cam_statis}
 		monitor_cam_empty
 	}
-        done
+	done
 }
 
 function monitor_cam_empty()
 {
-	(
-		cd ${run_dir}
-		sudo rm -r video*
-	)
+	local warn_threold=$(df -h|grep mmcblk |awk '{print $5}'|awk -F '%' '{print $1}')
+	if [ "${warn_threold}" -gt 90 ]
+	then
+	{
+		(
+			cd ${run_dir}
+			sudo rm -r video*
+		)
+	}
+	fi
 }
 
 function monitor_cam()
 {
 	local run_dir=/home/worker/camera
 	local config_json=${run_dir}/config.json
-        local log_date=$(date "+%Y-%m-%d-%H-%M-%S")
-        local log_dir=${run_dir}/log/check_camera/${log_date}
-        local log_summary=${run_dir}/check_camera.summary
-        [ ! -d "${log_dir}" ] && mkdir -p ${log_dir}
-        monitor_cam_num
+	local log_date=$(date "+%Y-%m-%d-%H-%M-%S")
+	local log_dir=${run_dir}/log/check_camera/${log_date}
+	local log_summary=${run_dir}/check_camera.summary
+	[ ! -d "${log_dir}" ] && mkdir -p ${log_dir}
+	monitor_cam_num
 	monitor_cam_empty
-        monitor_cam_excute &
-        monitor_cam_result
+	monitor_cam_excute &
+	monitor_cam_result
 }
 
 monitor_cam
